update tasks refactoring
