+++ src/qgis_geonode/tasks/tasks.py
@@ -1,10 +1,15 @@
 import typing
 import urllib.parse
+import shutil
+import tempfile
+from pathlib import Path
+import dataclasses
 
 from qgis.PyQt import (
     QtCore,
     QtWidgets,
     QtGui,
+    QtNetwork,
 )
 
 from ..apiclient import (
@@ -13,6 +18,15 @@ from ..apiclient import (
 )
 import qgis.core
 from ..utils import log
+from ..tasks import network_task
+from ..utils import log, sanitize_layer_name
+from .. import network
+
+
+@dataclasses.dataclass()
+class ExportFormat:
+    driver_name: str
+    file_extension: str
 
 
 class ThumbnailLoaderTask(qgis.core.QgsTask):
@@ -164,3 +178,228 @@ class LayerLoaderTask(qgis.core.QgsTask):
             params["authcfg"] = self.api_client.auth_config
         uri = " ".join(f"{key}='{value}'" for key, value in params.items())
         return qgis.core.QgsVectorLayer(uri, self.brief_dataset.title, "WFS")
+
+
+class LayerUploaderTask(network_task.NetworkRequestTask):
+    VECTOR_UPLOAD_FORMAT = ExportFormat("ESRI Shapefile", "shp")
+    RASTER_UPLOAD_FORMAT = ExportFormat("GTiff", "tif")
+
+    layer: qgis.core.QgsMapLayer
+    allow_public_access: bool
+    _upload_url: QtCore.QUrl
+    _temporary_directory: typing.Optional[Path]
+
+    def __init__(
+        self,
+        layer: qgis.core.QgsMapLayer,
+        upload_url: QtCore.QUrl,
+        allow_public_access: bool,
+        authcfg: str,
+        network_task_timeout: int,
+        description: str = "LayerUploaderTask",
+    ):
+        """Task to perform upload of QGIS layers to remote GeoNode servers."""
+        super().__init__(
+            requests_to_perform=[],
+            authcfg=authcfg,
+            description=description,
+            network_task_timeout=network_task_timeout,
+        )
+        self.response_contents = [None]
+        self.layer = layer
+        self.allow_public_access = allow_public_access
+        self._upload_url = upload_url
+        self._temporary_directory = None
+
+    def run(self) -> bool:
+        if self._is_layer_uploadable():
+            source_path = Path(
+                self.layer.dataProvider().dataSourceUri().partition("|")[0]
+            )
+            export_error = None
+        else:
+            log(
+                "Exporting layer to an uploadable format before proceeding with "
+                "the upload..."
+            )
+            source_path, export_error = self._export_layer_to_temp_dir()
+        log(f"source_path: {source_path}")
+        if export_error is None:
+            sld_path, sld_error = self._export_layer_style()
+            log(f"sld_path: {sld_path}")
+            if sld_path is None:
+                log(
+                    f"Could not export the layer's style as SLD "
+                    f"({sld_error}), skipping..."
+                )
+            multipart = self._prepare_multipart(source_path, sld_path=sld_path)
+            with network.wait_for_signal(
+                self._all_requests_finished, timeout=self.network_task_timeout
+            ) as event_loop_result:
+                request = QtNetwork.QNetworkRequest(self._upload_url)
+                request.setHeader(
+                    QtNetwork.QNetworkRequest.ContentTypeHeader,
+                    f"multipart/form-data; boundary={multipart.boundary().data().decode()}",
+                )
+                if self.authcfg:
+                    auth_manager = qgis.core.QgsApplication.authManager()
+                    auth_added, _ = auth_manager.updateNetworkRequest(
+                        request, self.authcfg
+                    )
+                else:
+                    auth_added = True
+                if auth_added:
+                    qt_reply = self._dispatch_request(
+                        request, network.HttpMethod.POST, multipart
+                    )
+                    multipart.setParent(qt_reply)
+                    request_id = qt_reply.property("requestId")
+                    self._pending_replies[request_id] = network.PendingReply(
+                        0, qt_reply, False
+                    )
+                else:
+                    self._all_requests_finished.emit()
+            loop_forcibly_ended = not bool(event_loop_result.result)
+            if loop_forcibly_ended:
+                result = False
+            else:
+                result = self._num_finished >= len(self.requests_to_perform)
+        else:
+            result = False
+        return result
+
+    def finished(self, result: bool) -> None:
+        if self._temporary_directory is not None:
+            shutil.rmtree(self._temporary_directory, ignore_errors=True)
+        super().finished(result)
+
+    def _prepare_multipart(
+        self, source_path: Path, sld_path: typing.Optional[Path] = None
+    ) -> QtNetwork.QHttpMultiPart:
+        main_file = QtCore.QFile(str(source_path))
+        main_file.open(QtCore.QIODevice.ReadOnly)
+        sidecar_files = []
+        if sld_path is not None:
+            sld_file = QtCore.QFile(str(sld_path))
+            sld_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("sld_file", sld_file))
+        if self.layer.type() == qgis.core.QgsMapLayerType.VectorLayer:
+            dbf_file = QtCore.QFile(str(source_path.parent / f"{source_path.stem}.dbf"))
+            dbf_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("dbf_file", dbf_file))
+            prj_file = QtCore.QFile(str(source_path.parent / f"{source_path.stem}.prj"))
+            prj_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("prj_file", prj_file))
+            shx_file = QtCore.QFile(str(source_path.parent / f"{source_path.stem}.shx"))
+            shx_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("shx_file", shx_file))
+        elif self.layer.type() == qgis.core.QgsMapLayerType.RasterLayer:
+            # when uploading tif files GeoNode seems to want the same file be uploaded
+            # twice - one under the `base_file` form field and another under the
+            # `tif_file` form field. This seems like a bug in GeoNode though
+            tif_file = QtCore.QFile(str(source_path))
+            tif_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("tif_file", tif_file))
+        permissions = {
+            "users": {},
+            "groups": {},
+        }
+        if self.allow_public_access:
+            permissions["users"]["AnonymousUser"] = [
+                "view_resourcebase",
+                "download_resourcebase",
+            ]
+        multipart = network.build_multipart(
+            self.layer.metadata(), permissions, main_file, sidecar_files=sidecar_files
+        )
+        # below we set all QFiles as children of the multipart object and later we
+        # also make the multipart object a children on the network reply object. This is
+        # done in order to ensure deletion of resources at the correct time, as
+        # recommended by the Qt documentation at:
+        # https://doc.qt.io/qt-5/qhttppart.html#details
+        main_file.setParent(multipart)
+        for _, qt_file in sidecar_files:
+            qt_file.setParent(multipart)
+        return multipart
+
+    def _is_layer_uploadable(self) -> bool:
+        """Check if the layer is in a format suitable for uploading to GeoNode."""
+        ds_uri = self.layer.dataProvider().dataSourceUri()
+        fragment = ds_uri.split("|")[0]
+        extension = fragment.rpartition(".")[-1]
+        return extension in (
+            self.VECTOR_UPLOAD_FORMAT.file_extension,
+            self.RASTER_UPLOAD_FORMAT.file_extension,
+        )
+
+    def _export_layer_to_temp_dir(
+        self,
+    ) -> typing.Tuple[typing.Optional[Path], typing.Optional[str]]:
+        if self._temporary_directory is None:
+            self._temporary_directory = Path(tempfile.mkdtemp(prefix="qgis_geonode_"))
+        if self.layer.type() == qgis.core.QgsMapLayerType.VectorLayer:
+            exported_path, error_message = self._export_vector_layer()
+        elif self.layer.type() == qgis.core.QgsMapLayerType.RasterLayer:
+            exported_path, export_error = self._export_raster_layer()
+            error_message = str(export_error)
+        else:
+            raise NotImplementedError()
+        return exported_path, (error_message or None)
+
+    def _export_vector_layer(
+        self,
+    ) -> typing.Tuple[typing.Optional[Path], str]:
+        sanitized_layer_name = sanitize_layer_name(self.layer.name())
+        target_path = self._temporary_directory / f"{sanitized_layer_name}.shp"
+        export_code, error_message = qgis.core.QgsVectorLayerExporter.exportLayer(
+            layer=self.layer,
+            uri=str(target_path),
+            providerKey="ogr",
+            destCRS=qgis.core.QgsCoordinateReferenceSystem(),
+            options={
+                "driverName": "ESRI Shapefile",
+            },
+        )
+        if export_code == qgis.core.Qgis.VectorExportResult.Success:
+            result = (target_path, error_message)
+        else:
+            result = (None, error_message)
+        return result
+
+    def _export_raster_layer(
+        self,
+    ) -> typing.Tuple[typing.Optional[Path], typing.Optional[int]]:
+        sanitized_layer_name = sanitize_layer_name(self.layer.name())
+        target_path = (
+            self._temporary_directory
+            / f"{sanitized_layer_name}.{self.RASTER_UPLOAD_FORMAT.file_extension}"
+        )
+        writer = qgis.core.QgsRasterFileWriter(str(target_path))
+        writer.setOutputFormat(self.RASTER_UPLOAD_FORMAT.driver_name)
+        pipe = self.layer.pipe()
+        raster_interface = self.layer.dataProvider()
+        write_error = writer.writeRaster(
+            pipe,
+            raster_interface.xSize(),
+            raster_interface.ySize(),
+            raster_interface.extent(),
+            raster_interface.crs(),
+            qgis.core.QgsCoordinateTransformContext(),
+        )
+        if write_error == qgis.core.QgsRasterFileWriter.NoError:
+            result = (target_path, None)
+        else:
+            result = (None, write_error)
+        return result
+
+    def _export_layer_style(self) -> typing.Tuple[typing.Optional[Path], str]:
+        sanitized_layer_name = sanitize_layer_name(self.layer.name())
+        if self._temporary_directory is None:
+            self._temporary_directory = Path(tempfile.mkdtemp(prefix="qgis_geonode_"))
+        target_path = self._temporary_directory / f"{sanitized_layer_name}.sld"
+        saved_sld_details, sld_exported_flag = self.layer.saveSldStyle(str(target_path))
+        if "created default style" in saved_sld_details.lower():
+            result = (target_path, "")
+        else:
+            result = (None, saved_sld_details)
+        return result
